Define sample AJAX request route
A script defining a JavaScript function is added to the individual quiz
view if the logged in user has Edit permission. This function sends an
XMLHttpRequest to a route exposed by the service, which is reserved for
users with Edit permission by Spring Security.

There are no triggers on the page for this function but it can be run
manually from the console. It has no effect on the back end other than
to produce an HTTP response based on whether appropriate parameters were
sent and whether the logged in user has appropriate permission to access
the route; the response status code is displayed in an alert and the
XMLHttpRequest object is printed in the console to aid development;