Enable creating database record from the front end
Using the createAnswer JavaScript function (which must still be executed
manually in the console) a request can be sent which triggers a call to
execute the appropriate database function.

Basic type checking has also been added to the JavaScript, to prevent
questions being created with the string "undefined" if no value is given
for the second parameter in the function call.